clock_angles/requirements.txt
Flask==1.1.2
pytest==5.3.2
google-cloud-datastore==1.12.0